requirements.txt
py-cord
pytube
ffmpeg-python